Strip origins for recursively vendored Pulumi dependencies
Govendor prefers to keep the recursively vendored dependency
directories as the "origin" for all dependency downloads.  This
is presumably for repeatability.  But, in our case, we aren't
checking in the vendor/ directory, so this fails.

We will strip out the "origin" for recursively vendored Pulumi
dependencies.  It seems superfluous to me anyway, since we have
both the SHA1 commit hash and a checksum of the contents.