Tolerate `nil` returns from `Apply`
At the moment, our code assumes the state object returned from
`Apply` is non-`nil`.  We have seen crashes that imply this isn't
the case and, indeed, the source code reveals that this can easily
happen:

https://github.com/hashicorp/terraform/blob/master/helper/schema/resource.go#L211

It's not immediately apparent to me the cause of that condition,
since (confusingly) it's a !d.RequiresNew that is preceded by an
if guard that d.RequiresNew; but there's plenty of mutable state
in here, and indeed L205 just above mutates the ID.  And notably,
the recordCurrentSchemaVersion function used liberally throughout
this code defends itself against nil states too:

https://github.com/hashicorp/terraform/blob/master/helper/schema/resource.go#L543

Perplexingly, pulumi/pulumi-terraform#56 happens indeterminitely.
And I'm surprised we haven't been hitting this all the time.

Nevertheless, it's clear that our handling of the state object
isn't correct, and so I'd like to submit this while concurrently
verifying that this has anything to do with the actual bug ...