Add back "-"; reduce random length to 7
This change adds back the "-" separator, and reduces the random
hash length from 8 to 7.  This increases the odds of collisions
slightly, but has the aesthetic benefit of being able to easily
see the user-specified part of the name versus the auto-generated
part of the name.  (TBH, I'm not terribly worried about collisions
these days, since most people in practice do things like prefix
with stack names, etc.  Even 7 is probably a little paranoid.)