Allow explicitly overriding MaxItemsOne in SchemaInfo
This allows a package to override the default behaviour of projecting MaxItems==1 as a scalar and MaxItems!=1 as an array.

For example, if the TF schema should have had a MaxItems==1 annotation (only one item can be supported by the underlying cloud platform API), then a package may want to force projection as a scalar.

And in cases where MaxItem==1 is set in the TF schema, but it is known that this is temporary (the underlying cloud platform is designed to support multiple, but currently only supports one), it may make sense to force projection as an array to ensure a later addition is not a breaking change.