Hide the private constructor overload used by `.get`
This is a targetted fix to address the regressed error checking in #153.

Ideally, we would go further, and avoid using the resource constructor as the place we do this - but this change at least removes publically acknowledging this as part of the supported API, so also provides us the ability to make such a change more easily in the future.

Fixes #153.