Don't emit Removed fields in sub-structure properties
tfbridge/gen checked the terraform schema "Removed" property for top
level fields in resources, but it wasn't checked for fields of
sub-structures (such as "networks" on an openstack compute_instance_v2
resource).

This adds a method to do a recursive clean on resources to remove all
"Removed" attributes. The method is used by both tfbridge and tfgen when
reading from the terraform providers ResourcesMap and DataSourcesMap.