Generate source diffs in Travis
When the code generation logic in pulumi-terraform changes, it would
be helpful to see the impact to downstream repositories. This change
does so by exploiting travis's build stages to run builds in all the
tfgen'd repositories which generate the code and either push it to
github (if the branch PR comes from a topic branch) or just runs "git
diff" into the build log.

`pulumi-bot` then adds an old style status check to the PR which links
you to the source diff.

Ideally we'd run tests, but doing so requires a fair bit more work,
because pulumi-terraform's environment does not have all the
environment variaibles nessecary to actually run tests in other
repositories and may not have a complete toolchain either (e.g. the
GCP SDK will not be installed and we won't have configured the
credentials).

I think we can address this by queuing builds against the downstream
repositories with the API itself and then managing the status checks
ourselves, but landing this as is gives a mesurable improvment to our
inner loop when working on the code generators here.