Project default values from literals or envvars.
It is common for provider properties in Terraform to have defaults that
are pulled from environment variables or a literal value. These changes
add support for this scenario by adding an additional property to the
schema overlay containing envvars for default values. This property is
then used in concert with the existing literal default to generate
appropriate code for each language.