Fix types and marshalling for Resource-Elem TypeMap values.
Terraform maps are odd, to say the least. In general a map value may be
either a list of string -> value maps or a string -> value map itself.
Maps that have an element type that is a *schema.Resource, however, must
be lists of string -> value maps.

These changes update the Pulumi -> TF config marshaller so that in the
latter case we appropriately wrap input maps in an array. The codegen
for Node and Go has also been updated to accurately reflect the type.

Fixes #231.