Relax asset reflection.
The TF <-> Pulumi marshaller attempts to turn outputs that corresponds
to inputs that were assets back in to the original asset value so that
the engine is able to provide better diffs. The logic it used to do this
was overly strict, and required that all properties that had asset
mappings have asset values. These changes relax that condition s.t.
values that are not assets are passed through as-is.

Fixes #139.