Emit correct presence checks for required props
Coercing a value to a boolean is not enough to determine whether a
property was provided in Python - it may have been provided, but it may
be a value that Python does not consider to be truthy, such as the
number zero.

This commit tests properties explicitly against None (their Python
default value) to check and see if they have been provided.