Add an `AutoNamed` field to DefaultInfo
And create a marshallable version of the DefaultInfo type.

The `AutoNamed` field will be true for a given default if the default
represents an auto-generated name. This information will be consumed by
`tf2pulumi` in order to decide whether or not to elide a particular
property value when name elision is enabled.

Conrtributes to #325.