Fall back to the schema type for strings
Terraform occasionally sends back strings for fields whose schema type
indicates that they are not strings. If this is the case, convert the
string to the requested primitive instead of returning the string
verbatim.

This fixes a variety of type errors that can occur when using resource
outputs in strongly typed languages.