Use Python name to refer to submodules in __all__
Prior to this commit we were using a module's Pulumi name in __all__,
which is supposed to list the full set of submodules contained within a
module. Unfortunately this did not correspond to the on-disk module
format, which is a Python name (camel-cased), and thus it was impossible
to import this module.

This commit fixes the issue by using the Python name to refer to
submodules when generating __all__.