Track default population in inputs
Add a reserved property to each map value present in an input bag that
tracks which keys in the map were populated using default values. When
performing default application, only use a value from an old input bag
if it was itself a default value.

This allows the following sequence of events:
- Create a resource with a value explicitly specified for property P
- Remove the value for property P and update the resource. The new value
  for property P will be whatever its default value is, if any.

Fixes https://github.com/pulumi/pulumi-aws/issues/439.