Bridge between PEP440 and semver versions
Python PEP440 is slightly different in incompatible ways from semver, so
sending our own PEP440 version to the engine that is expecting a semver
string caused problems when using prerelease builds.

To fix this, this commit makes sure that versions are explicitly coerced
into semver before handing them to the core Pulumi SDK. The engine
always expects semver strings, so converting to semver as early as
possible prevents odd incompatibilities that arise when PEP440 versions
are sent to the engine.