Use a schematized reader in MakeTerraformResult.
This prevents bugs where `flatmap.Expand` would misinterpret the
contents of a resource's state with respect to the shape described by
the resource's schema. The most critical misinterpretation involves
submaps with keys that contain "." characters: in this case, the key is
misinterpreted as describing another submap rather than a string ->
string entry. For example, given the following TF state:

```
"attributes": {
    "tags.%": "1",
    "tags.pulumi.io/foo/bar": "baz",
},
```

flatmap will expand this to:
```
{
    "tags": {
        "pulumi": {
            "io/foo/bar": "baz"
        }
    }
```

instead of the expected
```
    "tags": {
        "pulumi.io/foo/bar": "baz"
    }
```

Fixes #394.