Ignore state-only diffs.
It is possible for a Terraform diff to contain diffs to properties that
are only present in the state. These diffs typically exist because
`CustomizeDiff` has marked a computed property as differing (see e.g.
the `last_modified` property of `aws_lambda_function`). We do not care
about these diffs and can safely remove them from the detailed diff.