Fix bug for providers that don't have any nested types
Addresses two issues:

1.  Ensure the list of files are included in `tsconfig.json` for providers that don't have any nested types.

2. Avoid emitting `inputApi`/`outputApi` in modules that don't have any nested types.

The way we accomplish (2) isn't ideal. Since we're computing the nested types as we're generating each resource file (instead of precomputing them up-front, which would be the proper way of doing this), we won't know if we should include the `inputApi`/`outputApi` imports at the top of the file until we've written the file. So, at the end of emitting the file, if we hadn't written the imports and we now have nested types, regenerate the file again, this time with the imports at the top.