Set delete-before-replace for manually-named resources.
Determine if a resource is auto-named by checking its list of top-level
defaults for a property that was populated by a default value that is an
autoname. If the defaults list is missing, consider the resource
auto-named for better backwards compatibility.

These changes also fix a bug in the logic that determined the legacy
replaces list.

Fixes #383.