Change strategies.
The original code would have enabled DBR for any resource that is not
autonamed, which is subtly different than enabling DBR only for
resources that are manually-named. With those changes, any resource
that does not have an autonameable property would always have been
deleted before being replaced. This is probably not the correct default
for resources that do not have an autonamable property.

This commit changes the way we decide whether or not a resource's name
requires DBR: rather than looking for the presence of a default that
came from an autoname, the code now looks for an input property with a
manually-specified (i.e. non-default) value that _could_ have been
autonamed.