Fix nested property names for Python docstrings.
Nested property names in Python are remapped using the global
snake-to-camel-case and camel-to-snake-case tables emitted as part of
each package. Due to bugs in the code that generates these mappings,
most nested property names are _not_ mapped at all, and instead simply
use camel casing. These changes update the docstring generation for
nested properties to respect the appropriate mappings.