Remove Terraform Bridge components
Since the Terraform Plugin SDK was released, the Terraform Bridge has
no need to depend on Terraform itself - however the resource provider
which provides `RemoteStateReference` does still need the full
dependency. In order to avoid imposing this on providers, the bridge
code (without the provider) now lives in the `pulumi-terraform-bridge`
repository instead, so this commit removes the duplicate code.