Add a terraform.Module resource type
This adds a Module resource type that can provision a Terraform
module from within Pulumi. It is dynamically typed on its inputs and
outputs, and shells out to the CLI to perform the actual provisioning.

This is a very quick and dirty prototype and is not intended for
actual use, just exploration of some ideas.