Support YAML, Java and Registry by adapting `pulumi-go-provider` (#769)
This PR rewrites the `pulumi-terraform` provider on top of the
[pulumi-go-provider](https:www.github.com/pulumi/pulumi-go-provider)
SDK. The old provider hadn't been meaningfully updated in a long time
and had fallen far out of our definition of a "modern" Pulumi provider.
It was missing:

-   Support for Pulumi YAML (long since GA'ed)
-   Support for Java (GA'ed)
-   Documentation in the Pulumi Registry

This rewrite gives us generated Pulumi schemas that are known to match
the provider we ship with. That gives us high-quality and homogeneous
SDKs, since each SDK is generated from a modern version of codegen. It
also lets us add `pulumi-terraform` to the Pulumi Registry.


<a id="org45aa532"></a>

# State of the PR

This PR ships a modern, releasable and fully functional provider with
support for a subset of Terraform's state backends. There are tests in
every official Pulumi language: Pulumi YAML, typescript, python, C#, Go
and Java.

CI was generated from [ci-mgmt](https://github.com/pulumi/ci-mgmt)
(thanks @mjeffryes: [Move native providers to templates
#1480](https:https://www.github.com/pulumi/ci-mgmt/issues/1480)), and
then modified to be compatible with a `pulumi-go-provider` based
approach (no tfgen binary) and to add a `yaml` test tag.


<a id="orge1acfb9"></a>

# Backends

Terraform supports 15 different backends. This PR currently supports 2:
`local` and `remote`.

Supported backends:

- [ ] `artifactory`
- [ ] `azurerm`
- [ ] `consul`
- [ ] `etcdv2`
- [ ] `etcdv3`
- [ ] `gcp` (or `gcs`, depending on the SDK)
- [ ] `http`
- [X] `local`
- [ ] `manta`
- [ ] `oss`
- [ ] `postgres`
- [X] `remote`
- [ ] `s3`
- [ ] `swift`
- [ ] [Kubernets](https://github.com/pulumi/pulumi-terraform/issues/553)
(not supported in the current SDK)


<a id="org42cedf0"></a>

# Recommendation going forward

We should merge this PR. That would give us immediate benefits:

-   Start supporting some backends in all supported languages.
- Get a framework in place for adding new backends, including those not
supported in the current provider.
- Get `pulumi-terraform` into the Pulumi Registry, exposing high quality
docs to our
    customers (and Pulumi AI).

Engineers and open source contributors can easily add new backends as
they become necessary. We will open a tracking issue for each missing
backend. Each backend takes about 20 minutes to add, plus testing.
Getting an integration test takes the vast majority of the time.

When this PR is released, it will be a breaking change and needs to be
released as v6.0.0. Users who don't see a need to migrate or who are
using backends that are not yet supported can continue to use the
existing v5 version without interruption.

Fixes #710 
Fixes #702 
Fixes #633
Fixes #727

---------

Co-authored-by: Ian Wahbe <me@iwahbe.com>